+++ lib/rich_text_renderer/text_renderers/subscript_renderer.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+require_relative './base_inline_renderer'
+
+module RichTextRenderer
+  # subscript node renderer.
+  class SubscriptRenderer < BaseInlineRenderer
+    protected
+
+    def render_tag
+      'sub'
+    end
+  end
+end