+++ lib/rich_text_renderer/text_renderers/superscript_renderer.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+require_relative './base_inline_renderer'
+
+module RichTextRenderer
+  # superscript node renderer.
+  class SuperscriptRenderer < BaseInlineRenderer
+    protected
+
+    def render_tag
+      'sup'
+    end
+  end
+end